pyproject.toml
[project]
name = "film-analyze"
version = "0.1.0"
description = "CLIP-based video fragment search system"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "clip @ git+https://github.com/openai/CLIP.git",
    "opencv-python>=4.8.0",
    "scenedetect[opencv]>=0.6.0",
    "ffmpeg-python>=0.2.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "hydra-core>=1.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "open-clip-torch>=3.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src", "schemas", "conf"]

